Optimize TextDataset to use int32 for input tensors
- Changed `astype(np.int64)` to `astype(np.int32)` for input tensor `x` in `evaluate.py`.
- Preserved `int64` for target tensor `y` to satisfy `CrossEntropyLoss` requirements.
- Added `tests/test_evaluate_dataset.py` to verify correctness of data types and values.
- Reduces serialized data item size by approx 24% (17KB -> 13KB).

+++ evaluate.py
@@ -20,7 +20,7 @@ class TextDataset(Dataset):
         return len(self.data) - self.block_size
 
     def __getitem__(self, idx):
-        x = torch.from_numpy(self.data[idx : idx + self.block_size].astype(np.int64))
+        x = torch.from_numpy(self.data[idx : idx + self.block_size].astype(np.int32))
         y = torch.from_numpy(self.data[idx + 1 : idx + 1 + self.block_size].astype(np.int64))
         return x, y
